% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KX2006.R
\name{KX2006}
\alias{KX2006}
\title{Koenker and Xiao (2006) QAR(1) model}
\usage{
KX2006(
  Y,
  n.sims = 10000,
  n.thin = 10,
  n.burnin = 10000,
  n.report = 1000,
  inits = rep(0, 4),
  prior = c(10, 3, 3),
  tol = 1e-15
)
}
\arguments{
\item{Y}{MATRIX, rows are independent, columns have autoregression}

\item{n.sims, n.thin, n.burnin, n.report}{(i) Number of iterations after 
burn-in. (ii) Thinning rate. (iii) Number of iterations discarded at the 
beginning. (iv) Report the number of iterations rate.}

\item{inits}{Initial values (transformed scale \eqn{(-\infty,\infty)})}

\item{prior}{Vector (length 3). Standard deviation of the zero-mean normal 
prior for (1) \eqn{\mu}, (2) \eqn{\log \sigma}, and (3) 
\eqn{\log \gamma_1}, respectively. 
(\eqn{\gamma_0 \sim U(0,1)} by default.)}

\item{tol}{Tolerance in the univariate rootfinder. A value too small can 
lead to numerical overflow in the likelihood}
}
\value{
A \code{"KX2006"} list with elements:
  \item{params}{Matrix where rows are simulations and cols are parameters
  \deqn{\mu,\sigma,\gamma_0,\gamma_1}}
  \item{\code{y}}{Data fitted}
}
\description{
This function fits the QAR model by Koenker and Xiao (2006) in a Bayesian 
framework, i.e.,
\deqn{Q_{Y_{tl}}(\tau \mid y_{t,l-1}) = \mu + \sigma \Phi(\tau) + \min\{\gamma_0 + \gamma_1 \tau, 1\} y_{t,l-1}}
with \eqn{\gamma_0 \in (0,1)} and \eqn{\gamma_1 > 0}.
}
\references{
Koenker, R. and Xiao, Z. (2006). 
“Quantile autoregression.” 
\emph{Journal of the American Statistical Association}, \strong{101}(475): 980–990
}
