% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MQAR1.R
\name{MQAR1K1}
\alias{MQAR1K1}
\title{Multivariate QAR(1) with K = 1}
\usage{
MQAR1K1(
  Y,
  n.sims = 10000,
  n.thin = 10,
  n.burnin = 10000,
  n.report = 1000,
  inits = rep(0, 9),
  prior = 3,
  tol = 1e-11
)
}
\arguments{
\item{Y}{ARRAY, rows are independent, columns have autoregression, 
third dim (length 2) have copula dep (multivariate)}

\item{n.sims, n.thin, n.burnin, n.report}{(i) Number of iterations after 
burn-in. (ii) Thinning rate. (iii) Number of iterations discarded at the 
beginning. (iv) Report the number of iterations rate.}

\item{inits}{Initial values}

\item{prior}{Number. Standard deviation of the normal prior for \eqn{\log a}'s and
\eqn{\log b}'s}

\item{tol}{Tolerance in the univariate rootfinder. A value too small can lead 
to numerical overflow in the likelihood}
}
\value{
MATRIX, rows are simulations, columns are parameters
\deqn{a_1,b_1,a_2,b_2,a_3,b_3,a_4,b_4,\rho}
}
\description{
First series
\deqn{\eta_{1,1}(\tau) = F_{a_1,b_1}(\tau)}
\deqn{\eta_{2,1}(\tau) = F_{a_2,b_2}(\tau)}
Second series
\deqn{\eta_{1,2}(\tau) = F_{a_3,b_3}(\tau)}
\deqn{\eta_{2,2}(\tau) = F_{a_4,b_4}(\tau)}
}
