% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QAR1.R
\name{QAR1K}
\alias{QAR1K}
\title{QAR(1) with K in a Basis (fixed a's and b's)}
\usage{
QAR1K(
  Y,
  K = 8,
  coef = c(0.5, 0.5, 1, 1, 0.5, 2, 2, 0.5, 1, 16, 16, 1, 4, 8, 8, 4),
  n.sims = 10000,
  n.thin = 10,
  n.burnin = 10000,
  n.report = 1000,
  inits = rep(0, 14),
  prior = 3,
  tol = 1e-15
)
}
\arguments{
\item{Y}{MATRIX, rows are independent, columns have autoregression}

\item{K}{integer with the number of terms in the basis}

\item{coef}{VECTOR of length \eqn{2 \times K} with 
\eqn{(a_1,b_1,\ldots,a_K,b_K)}}

\item{n.sims, n.thin, n.burnin, n.report}{(i) Number of iterations after 
burn-in. (ii) Thinning rate. (iii) Number of iterations discarded at the 
beginning. (iv) Report the number of iterations rate.}

\item{inits}{Initial values (transformed scale \eqn{(-\infty,\infty)})}

\item{prior}{Number. Standard deviation of the zero-mean normal prior for 
the \eqn{\phi}'s 
(the logistic transformation of the \eqn{\lambda}'s)}

\item{tol}{Tolerance in the univariate rootfinder. A value too small can lead 
to numerical overflow in the likelihood}
}
\value{
A \code{"QAR1K1"} list with elements:
  \item{params}{Matrix where rows are simulations and cols are parameters
  \deqn{\lambda_1^{(1)},\ldots,\lambda_{K-1}^{(1)},
        \lambda_1^{(2)},\ldots,\lambda_{K-1}^{(2)}}}
  \item{\code{y}}{Data fitted}
  \item{\code{coef}}{Coefficients \eqn{(a_1,b_1,\ldots,a_K,b_K)}}
}
\description{
This function fits the model 
\deqn{Q_{Y_{tl}}(\tau \mid y_{t,l-1}) = y_{t,l-1} \eta_1(\tau) + (1 - y_{t,l-1}) \eta_2(\tau)}
where for \eqn{j=1,2},
\deqn{\eta_j(\tau) = \sum_{k=1}^{K} \lambda_k^{(j)} F(\tau \mid a_k,b_k)}
with \eqn{\lambda_K^{(j)} = 1 - \sum_{k=1}^{K-1} \lambda_k^{(j)}},
\eqn{(a_k,b_k)_{k=1}^{K}} fixed in the arguments, and \eqn{F(\tau \mid a,b)}
the Kumaraswamy cdf with parameters \eqn{a} and \eqn{b}
}
