% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale01.R
\name{scale01}
\alias{scale01}
\title{Scale in \eqn{(0,1)}}
\usage{
scale01(y)
}
\arguments{
\item{y}{VECTOR or MATRIX to scale between 0 and 1}
}
\value{
VECTOR or MATRIX with values between 0 and 1
}
\description{
This function scales a variable in \eqn{(0,1)} following ideas from the 
theory of order statistics in uniform random variables
\deqn{y_{t} = \frac{y_t^* - m}{M - m},}
where \eqn{m = (T y_{(1)}^{*} - y_{(T)}^{*})/(T-1)} and 
\eqn{M = (T y_{(T)}^{*} - y_{(1)}^{*})/(T-1)}.
}
