% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQAR1.R
\name{SGPQAR1K1}
\alias{SGPQAR1K1}
\title{Spatial GP QAR(1) with K = 1}
\usage{
SGPQAR1K1(
  Y,
  coords,
  n.sims = 1e+06,
  n.thin = 1000,
  n.burnin = 1e+05,
  n.report = 1000,
  inits = rep(0, 9 + 4 * dim(Y)[3]),
  prior = c(3, 3),
  tol = 1e-15
)
}
\arguments{
\item{Y}{ARRAY, rows are independent, columns have autoregression, 
third dim have copula dep (space)}

\item{coords}{COORDINATES \eqn{n \times 2} of the locations}

\item{n.sims, n.thin, n.burnin, n.report}{(i) Number of iterations after 
burn-in. (ii) Thinning rate. (iii) Number of iterations discarded at the 
beginning. (iv) Report the number of iterations rate.}

\item{inits}{Initial values}

\item{prior}{Vector (length 2). Standard deviation of the normal prior 
for (1) \eqn{a}'s and \eqn{b}'s, and (2)
\eqn{\log \sigma}'s, respectively}

\item{tol}{Tolerance in the univariate rootfinder. A value too small can lead 
to numerical overflow in the likelihood}
}
\value{
MATRIX, rows are simulations, columns are parameters
  \deqn{\gamma,\exp a_1,\exp b_1,\exp a_2,\exp b_2,
  \sigma_{a_1}^2,\sigma_{b_1}^2,\sigma_{a_2}^2,\sigma_{b_2}^2,
  a_1(s_1),\ldots,a_1(s_n),b_1(s_1),\ldots,a_2(s_1),\ldots,b_2(s_n)}
}
\description{
\deqn{\eta_1(s;\tau) = F_{a_1(s),b_1(s)}(\tau)}
\deqn{\eta_2(s;\tau) = F_{a_2(s),b_2(s)}(\tau)}

Burn-in of 100,000 iterations. Run at least 1,000,000 n.sims (might take 8/9 hours)
}
