% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQAR1.R
\name{SQAR1K1}
\alias{SQAR1K1}
\title{Spatial QAR(1) with K = 1}
\usage{
SQAR1K1(
  Y,
  coords,
  n.sims = 10000,
  n.thin = 10,
  n.burnin = 10000,
  n.report = 1000,
  inits = rep(0, 5),
  prior = 3,
  tol = 1e-15
)
}
\arguments{
\item{Y}{ARRAY, rows are independent, columns have autoregression, 
third dim have copula dep (space)}

\item{coords}{COORDINATES \eqn{n \times 2} of the locations}

\item{n.sims, n.thin, n.burnin, n.report}{(i) Number of iterations after 
burn-in. (ii) Thinning rate. (iii) Number of iterations discarded at the 
beginning. (iv) Report the number of iterations rate.}

\item{inits}{Initial values}

\item{prior}{Number. Standard deviation of the normal prior for \eqn{\log a}'s and
\eqn{\log b}'s}

\item{tol}{Tolerance in the univariate rootfinder. A value too small can lead 
to numerical overflow in the likelihood}
}
\value{
MATRIX, rows are simulations, columns are parameters
\deqn{a_1,b_1,a_2,b_2,\gamma}
}
\description{
\deqn{\eta_1(\tau) = F_{a_1,b_1}(\tau)}
\deqn{\eta_2(\tau) = F_{a_2,b_2}(\tau)}

Fast convergence
}
